pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "notes-creator"
version = "0.1.0"
description = "A desktop app to generate notes from YouTube video transcripts."
authors = [{ name = "Cline", email = "cline@example.com" }]
requires-python = ">=3.11"
dependencies = [
    "customtkinter",
    "google-genai",
    "youtube-transcript-api",
    "python-dotenv",
    "pydantic",
    "openai",
    "anthropic",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.uv]
dev-dependencies = [
    "pytest",
]